+++ R/mirai.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2022-2023 Hibiki AI Limited <info@hibiki-ai.com>
+# Copyright (C) 2022-2024 Hibiki AI Limited <info@hibiki-ai.com>
 #
 # This file is part of mirai.
 #
@@ -423,7 +423,7 @@ is_error_value <- is_error_value
 #'
 print.mirai <- function(x, ...) {
 
-  cat("< mirai >\n - $data for evaluated result\n", file = stdout())
+  cat("< mirai | $data >\n", file = stdout())
   invisible(x)
 
 }