+++ R/promises.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2023 Hibiki AI Limited <info@hibiki-ai.com>
+# Copyright (C) 2023-2024 Hibiki AI Limited <info@hibiki-ai.com>
 #
 # This file is part of mirai.
 #
@@ -53,10 +53,9 @@
 as.promise.mirai <- function(x)
   promises::promise(
     function(resolve, reject) {
-      later <- .getNamespace("later")[["later"]]
       query <- function()
         if (unresolved(x))
-          later(query, delay = 0.1) else
+          .[["later"]](query, delay = 0.1) else
             if (is_error_value(value <- .subset2(x, "value"))) reject(value) else resolve(value)
       query()
     }